supply correction sheet4/6
supply correction function merged into sheet4 functions
+ unnecessary paths removed from pahths.py
